AddNewWish, AddBook fixed, DetailsPage

+++ DigitalLibrary/BusinessLogic/BusinessLogicLayer.cs
@@ -1,6 +1,7 @@
 ﻿using DigitalLibrary.Models;
 using DigitalLibrary.Models.ViewModels;
 using DigitalLibrary.Repository.IRepository;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace DigitalLibrary.BusinessLogic
 {
@@ -81,17 +82,24 @@ namespace DigitalLibrary.BusinessLogic
         {
             string wwwRootPath = _webHostEnvironment.WebRootPath;
 
-            var oldImagePath = Path.Combine(wwwRootPath, Object.Book.ImageURL.TrimStart('\\'));
-            if (File.Exists(oldImagePath))
+            if (fileImage != null)
             {
-                File.Delete(oldImagePath);
+                var oldImagePath = Path.Combine(wwwRootPath, Object.Book.ImageURL.TrimStart('\\'));
+                if (File.Exists(oldImagePath))
+                {
+                    File.Delete(oldImagePath);
+                }
             }
 
-            var oldBookPath = Path.Combine(wwwRootPath, Object.Book.BookURL.TrimStart('\\'));
-            if (File.Exists(oldBookPath))
+            if(fileEbook != null)
             {
-                File.Delete(oldBookPath);
+                var oldBookPath = Path.Combine(wwwRootPath, Object.Book.BookURL.TrimStart('\\'));
+                if (File.Exists(oldBookPath))
+                {
+                    File.Delete(oldBookPath);
+                }
             }
+            
 
             _unitOfWork.Book.Remove(Object.Book);
             _unitOfWork.Save();
@@ -115,91 +123,6 @@ namespace DigitalLibrary.BusinessLogic
 
         #endregion
 
-        #region HOME
-        #endregion
 
-        #region WISHLIST
-
-        public void UpsertWish (BookVM Object, IFormFile fileImage, IFormFile fileEbook)
-        {
-            string wwwRootPath = _webHostEnvironment.WebRootPath;
-
-            if (fileImage != null)
-            {
-                string FileName = Guid.NewGuid().ToString();
-                var UploadImage = Path.Combine(wwwRootPath, @"images");
-                var Extension = Path.GetExtension(fileImage.FileName);
-
-                if (Object.Book.ImageURL != null)
-                {
-                    var oldImagePath = Path.Combine(wwwRootPath, Object.Book.ImageURL.TrimStart('\\'));
-                    if (System.IO.File.Exists(oldImagePath))
-                    {
-                        System.IO.File.Delete(oldImagePath);
-                    }
-                }
-
-                using (var fileStreamImage = new FileStream(Path.Combine(UploadImage, FileName + Extension), FileMode.Create))
-                {
-                    fileImage.CopyTo(fileStreamImage);
-                }
-                Object.Book.ImageURL = @"\images\" + FileName + Extension;
-            }
-
-            if (fileEbook != null)
-            {
-                string FileName = Guid.NewGuid().ToString();
-                var UploadEbook = Path.Combine(wwwRootPath, @"e-books");
-                var Extension = Path.GetExtension(fileEbook.FileName);
-
-                if (Object.Book.BookURL != null)
-                {
-                    var oldBookPath = Path.Combine(wwwRootPath, Object.Book.BookURL.TrimStart('\\'));
-                    if (System.IO.File.Exists(oldBookPath))
-                    {
-                        System.IO.File.Delete(oldBookPath);
-                    }
-                }
-
-                using (var fileStreamEbook = new FileStream(Path.Combine(UploadEbook, FileName + Extension), FileMode.Create))
-                {
-                    fileEbook.CopyTo(fileStreamEbook);
-                }
-                Object.Book.BookURL = @"\e-books\" + FileName + Extension;
-            }
-
-            if (Object.Book.Id == 0)
-            {
-                _unitOfWork.Book.Add(Object.Book);
-            }
-            else
-            {
-                _unitOfWork.Book.Update(Object.Book);
-            }
-
-            _unitOfWork.Save();
-        }
-
-        public void DeleteWish (BookVM Object, IFormFile fileImage, IFormFile fileEbook)
-        {
-            string wwwRootPath = _webHostEnvironment.WebRootPath;
-
-            var oldImagePath = Path.Combine(wwwRootPath, Object.Book.ImageURL.TrimStart('\\'));
-            if (System.IO.File.Exists(oldImagePath))
-            {
-                System.IO.File.Delete(oldImagePath);
-            }
-
-            var oldBookPath = Path.Combine(wwwRootPath, Object.Book.BookURL.TrimStart('\\'));
-            if (System.IO.File.Exists(oldBookPath))
-            {
-                System.IO.File.Delete(oldBookPath);
-            }
-
-            _unitOfWork.Book.Remove(Object.Book);
-            _unitOfWork.Save();
-        }
-
-        #endregion
     }
 }
